backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
sqlalchemy==1.4.23
psycopg2-binary==2.9.1
python-jose[cryptography]==3.3.0
PyJWT==2.1.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.5
aiofiles==0.7.0
python-dotenv==0.19.0
websockets>=10.4
openai==0.27.0
pydantic==1.10.12
email-validator==1.1.3
alembic==1.7.1
tenacity==8.0.1
httpx==0.18.2
requests==2.31.0
retry==0.9.2
fastapi-health==0.4.0
moviepy==1.0.3
Pillow==10.1.0
numpy==1.26.2
